src: server: huge amounts of fixes
Actually had a very heavy tests day, that brought this to become a little bit more stable, at least. Good night.
PS: investigations needed on cipher.c module source, as closing input(/_lock) and output file descriptors, actually closes stdin. Actually this is a huge problem, as when server accept connection, once this one requests to cipher a file, if it's the first one, no problem; otherwise, because of the already mentioned problem, the connected client waits forever for an answer about the exit status of the command, but the server is no more able to answer, as it thinks the file descriptor associated to the socket is actually 0 (as in the first connection, stdin was closed). Something bad, trust me, bad as the thousands hours I spent over all of this just to catch what the problem was.